Conflicts:
	libs/openFrameworks/.project
	libs/openFrameworks/graphics/ofGraphics.cpp
	libs/openFrameworks/graphics/ofGraphics.h
	libs/openFrameworks/graphics/ofShape.cpp
	libs/openFrameworks/graphics/ofShape.h
	libs/openFrameworksCompiled/project/osx/openFrameworksLib.xcodeproj/project.pbxproj

+++ libs/openFrameworks/3d/ofVertexData.cpp
@@ -355,7 +355,7 @@ void ofVertexData::setIndex(int i, int val){
 void ofVertexData::setupIndicesAuto(){
 	bIndicesChanged = true;
 	indices.clear();
-	for(int i = 0; i < vertices.size();i++){
+	for(int i = 0; i < (int)vertices.size();i++){
 		indices.push_back((GLuint)i);
 	}
 }
